requirements.txt
# Yo Mama Bot - Requirements
# Python 3.10+

# Core dependencies
google-genai==1.52.0
python-dotenv==1.2.1
doppler-sdk==1.3.0

# Platform integrations - Discord and Matrix
discord.py==2.6.4
matrix-nio==0.25.2

# Optional: Additional platforms
# atproto==0.0.63  # Bluesky
# Mastodon.py==2.1.4  # Mastodon

# Optional: Additional secret managers
# boto3>=1.26.0  # For AWS Secrets Manager
# hvac>=1.2.0    # For HashiCorp Vault
